Refactor map wrappers with HOC
